+++ src/minimal_harness/agent/__init__.py
@@ -1,9 +1,8 @@
-from .driver import DefaultAgentDriverFactory, RemoteAgentDriver, SSEAgentDriver
+from .driver import RemoteAgentDriver, RemoteAgentDriverFactory
 from .factory import AgentFactory, DefaultAgentFactory, LocalAgentFactory
 from .protocol import Agent, InputContentConversionFunction
 from .registry import AgentRegistry
 from .remote import RemoteAgent
-from .runner import SSEAgentRunner
 from .runtime import AgentRuntime, AgentRuntimeProtocol
 from .simple import SimpleAgent
 
@@ -14,12 +13,10 @@ __all__ = [
     "AgentRuntime",
     "AgentRuntimeProtocol",
     "DefaultAgentFactory",
-    "DefaultAgentDriverFactory",
     "InputContentConversionFunction",
     "LocalAgentFactory",
     "RemoteAgent",
     "RemoteAgentDriver",
-    "SSEAgentDriver",
-    "SSEAgentRunner",
+    "RemoteAgentDriverFactory",
     "SimpleAgent",
 ]